requirements.txt
Pillow==8.1.2
pyTelegramBotAPI==3.7.6
requests
python-decouple